Open Glaxnimate files (#24)
* fix(cmake): put glax_core before context

context uses glaxnimate code to get the current Document, so it must
come before it.

* feat(potrace): add GLAXNIMATE_SYSTEM_POTRACE option

* feat(context): add currentProjectFile to GlobalContext

* fix(context): import glax_core by proper include dir

we were importing with glax_core/ (include dir: src) instead of relying
on glax_core's include

* feat(glax): use DeclareModuleSetup for glaxnimate cmake

* fix(glax_core): do not hardcode module name

* docs(glax_core): format file

* fix(cmake): weird error from PCH

* fix(glax_core): we were linking to Qt twice

We were linking glax_core to Qt twice (once in DeclareModuleSetup.cmake
and another time in MODULE_LINK) so we were getting errors all around.
The error is to add a MODULE_LINK_QT flag to add the option to avoid
linking with Qt.

* fix(glax_core): link deps publicly

* fix(context): was not inheriting include directories from app::glax_core

we were linking with the module privately, so include directories didnt
contagiate

* fix(app): link libraries publicly so include directories propagate

* fix(glax_core): use absolute path for MODULE_INCLUDE

* revert: link glax_core publicly

* fix: set MODULE_ROOT not MODULE_INCLUDE (they have different effects)

* fix: include glax_core/model/document.hpp

include glax_core/... instead of model/document.hpp because context
doesn't have glax_core as an include directory

* revert(glax_core): link back privately

* fix(glax_core): Disable UNITY BUILD!!!

By disabling Unity build for glax_core, we got rid of most of the errors
we were getting!!!!

* fix(context): re-add glaxnimate::model::Document

* feat(appshell): implement file-open and file-import actions

* fix(context): not use Qt Signals for iglobalcontext

* fix(appshell): crappy fix, set include paths

* feat(projectscene): move EditPage to projectscene

* fix(projectscene): import issues

* fix: undo crappy fix, link libraries publicly so include dirs get
contagiated

* feat(projectscene): Add New Project dialog

* feat: menu shortcuts

* feat(appshell): tweak app menu

* feat(projectscene): implement project-open action

- created IProjectFilesController

* fix: typos

* fix(projectscene): make project files controller an Imoduleexport

* feat(projectscene): implement QFileDialog